requirements.txt
gradio
google-cloud-bigquery[pandas]
uvicorn
fastapi
debugpy